feat : Add TextComponent

+++ Despots_Project/Util/Image2D.cpp
@@ -34,7 +34,7 @@ void Image2D::Render(RECT rect)
 void Image2D::FrameRender(RECT rect, RECT frame)
 {
     m_image_rect = { (FLOAT)rect.left, (FLOAT)rect.top, (FLOAT)rect.right, (FLOAT)rect.bottom };
-    D2D1_RECT_F frameRect = { frame.left, frame.top, frame.right, frame.bottom };
+    D2D1_RECT_F frameRect = { (FLOAT)frame.left, (FLOAT)frame.top, (FLOAT)frame.right, (FLOAT)frame.bottom };
     if (mp_bitmap != NULL)
     {
         ImageManagerD2::GetInstance()->GetRenderTarget()->DrawBitmap(mp_bitmap, m_image_rect, 1, D2D1_BITMAP_INTERPOLATION_MODE_LINEAR, frameRect);
